webapp/requirements.txt
Flask==2.3.3
azure-eventhub==5.11.6
python-dotenv==1.0.0
